+++ src/enemies/bulk_enemy.hpp
@@ -5,7 +5,7 @@
 class BulkEnemy : public Enemy {
 
     public:
-        BulkEnemy(const TextureHolder &textures, float difficultyLevel = 1.f, float travelledDistance = 0.f, int directionIndex = 0);
+        BulkEnemy(const TextureHolder &textures, unsigned int difficultyLevel = 1, float travelledDistance = 0.f, int directionIndex = 0);
         void TakeHit(int damage, unsigned int bulletType);
 
     private:
