+++ src/enemies/multi_enemy.cpp
@@ -8,7 +8,7 @@
 
 
 //Making a leaf type enemy as a test of derived class
-MultiEnemy::MultiEnemy(const TextureHolder& textures, float difficultyLevel, float travelledDistance, int directionIndex, int spawnCount)
+MultiEnemy::MultiEnemy(const TextureHolder& textures, unsigned int difficultyLevel, float travelledDistance, int directionIndex, int spawnCount)
     : Enemy(Enemy::Multiplying, textures, difficultyLevel, travelledDistance, directionIndex), spawnTimer_(sf::Time::Zero), spawnCount_(spawnCount)
     { 
         showDeathAnimation_ = false;
